Tasks - Technical requirements

+++ Tasks/src/app/app.component.ts
@@ -1,23 +1,31 @@
 import { TasksCrudService } from './services/tasks.crud.service'; // Servicio
-import { Component, OnInit, ElementRef, Renderer2 } from '@angular/core';
+import { Component, OnInit, OnChanges, SimpleChanges, DoCheck, AfterViewInit, OnDestroy, Input} from '@angular/core'; // Lifecycle hooks
+import { ElementRef, Renderer2 } from '@angular/core'; // Gestión del DOM
 import { CommonModule } from '@angular/common'; // Necesario para implementar angular en el HTML
-import { ReactiveFormsModule } from '@angular/forms'; // Formularios reactivos
-import { FormGroup, FormControl } from '@angular/forms';
-
+import { ReactiveFormsModule, FormGroup, FormControl, FormsModule } from '@angular/forms'; // Formularios reactivos
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [CommonModule, ReactiveFormsModule],
+  imports: [CommonModule, ReactiveFormsModule, FormsModule],
   templateUrl: './app.component.html',
   styleUrl: './app.component.css'
 })
-export class AppComponent implements OnInit{
+export class AppComponent implements OnInit ,/*OnChanges,*/ DoCheck, AfterViewInit, OnDestroy {
+
+  private subscriptions: Subscription = new Subscription();
+  private timeoutId: any;
 
   tasks : any[] = [];
   form : FormGroup;
   updateTime : boolean = false;
   updateId : number = 0;
+  //@Input() filterValue: string = 'all';  // Para detectar cambios con ngOnChanges
+  filterValue: string = 'all';
+  alertModal : boolean = false;
+  shownModal : boolean = false;
+  errorModal : boolean = false;
 
   constructor(private tasksCrudService: TasksCrudService, 
               private renderer: Renderer2,
@@ -31,16 +39,56 @@ export class AppComponent implements OnInit{
     });
   }
 
+  // Lifecycle hooks
+  //  Init
   ngOnInit() {
     this.readTasks();
+  }
+  //  Check
+  ngDoCheck(): void {
+    this.tasksCrudService.CheckIfAllTasksCompleted().subscribe({
+      next: (allCompleted) => {
+        if (allCompleted &&!this.shownModal) {
+          this.alertModal = true;
+          this.shownModal = true;
+        }
+      }
+    });
+  }
+  // AfterView
+  ngAfterViewInit(): void {
+    const createTasksBttn = this.elementRef.nativeElement.querySelector('#createTaskButton');
+    const filterTasksSlct =  this.elementRef.nativeElement.querySelector('#taskFilter');
+    setTimeout(() => { // Evitar que el usuario cree o filtre tareas mientras se cargan
+      createTasksBttn.removeAttribute('disabled');
+      filterTasksSlct.removeAttribute('disabled');
+    }, 2000)
+  }
+  // Destroy
+  ngOnDestroy(): void {
+    localStorage.setItem('task', JSON.stringify(this.tasks)); // Guardar cambios no guardados
+    this.subscriptions.unsubscribe(); // Cancelar subscripciones
+    if (this.timeoutId) { // Limpiar setTimeout
+      clearTimeout(this.timeoutId);
+    }
+  }
+  //  Change
+  onFilterChange() {
+    this.filterTasks(this.filterValue);
   }
 
   //Create n Update
   createTask(): void{
-    const { complete, title, priority, dueDate, description } = this.form.value;
-    const id = (this.updateTime) ? this.updateId : parseInt(this.generateId());
-    this.tasksCrudService.createTask(id, complete, title, priority, dueDate, description, this.updateTime).subscribe({});
-    this.readTasks();
+    const { title, priority, dueDate, description } = this.form.value;
+    if([title, priority, dueDate, description].every(field => field !== null && field !== '')){
+      const id = (this.updateTime) ? this.updateId : parseInt(this.generateId());
+      const complete = false;
+      this.tasksCrudService.createTask(id, complete, title, priority, dueDate, description, this.updateTime).subscribe({});
+      this.readTasks();
+      this.shownModal = false;
+    }else{
+      this.errorModal = true;
+    }
   }
 
   //Read
@@ -50,15 +98,17 @@ export class AppComponent implements OnInit{
         this.tasks = tasks;
       },
       error: (error) => {
-        console.log(error);
+        console.error(error);
       }
     });
   }
 
   //Delete
   deleteTask(id : number): void{
-    this.tasksCrudService.deleteTask(id).subscribe({});
-    this.readTasks();
+    if(window.confirm("Are you sure you want to delete this task?")){
+      this.tasksCrudService.deleteTask(id).subscribe({});
+      this.readTasks();
+    }
   }
 
   generateId(): string{
@@ -98,49 +148,36 @@ export class AppComponent implements OnInit{
           this.renderer.addClass(element, 'completed');
           this.renderer.addClass(element, 'bg-secondary');
           this.tasksCrudService.createTask(task.id, !task.complete, task.title, task.priority, task.dueDate, task.description, true);
+          this.shownModal = false;
         }else{
           this.renderer.removeClass(element, 'completed');
           this.renderer.removeClass(element, 'bg-secondary');
           this.renderer.addClass(element, 'uncompleted');
           this.tasksCrudService.createTask(task.id, !task.complete, task.title, task.priority, task.dueDate, task.description, true);
+          this.alertModal = false;
         }
       }
     });
   }
 
   //Filter
-  filterTasks(event: any): void{
-    const filter = event.target.value;
-    if(filter == "all"){
+  filterTasks(filterEvent: string): void {
+  //filterTasks(event: any): void {
+    //const filterEvent = event.target.value;
+    if(filterEvent == "all"){
       this.readTasks();
-    }else if(filter == "complete"){
+    }else if(filterEvent == "complete"){
       this.readTasks();
       this.tasks = this.tasks.filter(task => task.complete);
-    }else if(filter == "pending"){
+    }else if(filterEvent == "pending"){
       this.readTasks();
       this.tasks = this.tasks.filter(task => !task.complete);
     }
   }
 
-}
+  closeModal(): void{
+    this.alertModal = false;
+    this.errorModal = false;
+  }
 
-/*
-//Filter
-  filterTasks(event: any): void{
-    const filter = event.target.value;
-    
-    this.tasksCrudService.readTasks().subscribe({
-    next: (tasks) => {
-      if (filter === "all") {
-        this.tasks = tasks;
-      } else if (filter === "complete") {
-        this.tasks = tasks.filter(task => task.complete);
-      } else if (filter === "pending") {
-        this.tasks = tasks.filter(task => !task.complete);
-      }
-    },
-    error: (error) => {
-      console.error('Error al filtrar las tareas:', error);
-    }
-  });
-*/
+}
